Initial commit
This is the result of the two first chapters, in which I made the Nodes
that will aid the process of parsing from Markdown to HTML by providing
a representation of the text in our code. Also wrote unit tests for the
Nodes to cover most cases I could think of (can't say I spent an
exaggerated amount of time on this, could probably use better coverage).